Move XIVDB module into its own folder and fix class/job names.
