+++ src/Lodestone/Modules/XIVDB/XIVDB.php
@@ -1,6 +1,8 @@
 <?php
 
-namespace Lodestone\Modules;
+namespace Lodestone\Modules\XIVDB;
+
+use Lodestone\Modules\HttpRequest;
 
 /**
  * Class XIVDB
@@ -9,8 +11,9 @@ namespace Lodestone\Modules;
  */
 class XIVDB
 {
-    use XIVDBApiTrait;
-    use XIVDBDataTrait;
+    use ApiTrait;
+    use DataTrait;
+    use ConvertTrait;
 
     const MAX_LEVEL = 60;
     const HOST = 'https://api.xivdb.com';
@@ -20,6 +23,7 @@ class XIVDB
     /** @var HttpRequest */
     private $http;
 
+    /** @var array $data */
     protected static $data = [];
 
     /**
