+++ src/Lodestone/Modules/XIVDB/data.php
@@ -12552,7 +12552,7 @@ return array (
     '9dba9727' => 24,
     56847897 => 25,
     '1d1ccece' => 26,
-    '39762c71' => 27,
+    'd1d1be2a' => 27,
     '4cb7a013' => 28,
     '932cf371' => 29,
     '42be4384' => 30,
@@ -12972,5 +12972,5 @@ return array (
     'ef2fbe38' => 128,
     'eeef4be4' => 130,
   ),
-  '_CACHED_' => 1497989058,
+  '_CACHED_' => 1498490958,
 );
