+++ src/Lodestone/Parser/Lodestone.php
@@ -3,7 +3,7 @@
 namespace Lodestone\Parser;
 
 use Lodestone\Modules\Routes;
-use Lodestone\Modules\XIVDB;
+use Lodestone\Modules\XIVDB\XIVDB;
 use Lodestone\Parser\Html\ParserHelper;
 
 /**
